refactor(build): move bundle scripts in the separate file
