+++ scripts/gulp/tasks/bundle/index.ts
@@ -1,0 +1,1 @@
+import './bundle';
